Refactor to use new test base for garbage collector
